requirements.txt
google-generativeai==0.3.2
click==8.1.7
rich==13.7.0
pathspec==0.12.1
python-dotenv==1.0.0
pydantic==2.5.2
asyncio==3.4.3
aiofiles==23.2.1
tabulate==0.9.0
subprocess32==3.5.4
